requirements.txt
numpy==1.19.5
pandas==1.1.5
scipy==1.5.4
scikit-learn==0.24.2
kneed-0.7.0
seaborn==0.11.2